MaterialSwitch and Instantiated objects implemented
I have now managed to integrate a material switch that would work for
objects with multiple materials (through element serial) and adapted the
array switch so that the trigger can work for different types of objects
(different arrays). I have added materials and props from the standard
unreal marketstore until we get assets to integrate in the environment.